azure-function/requirements.txt
azure-functions==1.17.0
requests==2.31.0
azure-storage-blob==12.19.0 